% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{example.comparisons}
\alias{example.comparisons}
\title{Example contrast matrix for input into the groupComparisonTMTPTM function}
\format{
A data frame with 9 rows and 6 variables.
}
\usage{
example.comparisons
}
\description{
Manually specified comparisons of interest for contrast.matrix
arguement of groupComparisonTMTPTM.
}
\details{
\itemize{
\item Condition_1, ... Condition_6 : Column names are conditions in dataset
\item 1-4, ... 5-6 : Row names are comparisons of interest
}
}
\examples{
head(example.comparisons)

}
\keyword{datasets}
