% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{quant.msstats.ptm}
\alias{quant.msstats.ptm}
\title{Example of output from proteinSummarizaiton function for PTM data}
\format{
A data frame with 19205 rows and 8 variables.
}
\usage{
quant.msstats.ptm
}
\description{
It is made from \code{\link{raw.ptm}}.
It is the output of proteinSummarization function from MSstatsTMT.
It should include the required columns as below.
The variables are as follows:
}
\details{
\itemize{
\item Run : MS run ID
\item Protein : Protein ID with modification site mapped in. Ex. Protein_1002_S836
\item Abundance: Protein-level summarized abundance
\item Channel : Labeling information (126, ... 131)
\item Condition : Condition (ex. Healthy, Cancer, Time0)
\item BioReplicate : Unique ID for biological subject.
\item TechRepMixture : Unique ID for technical replicate of one TMT mixture.
\item Mixture : Unique ID for TMT mixture.
}
}
\examples{
head(quant.msstats.ptm)

}
\keyword{datasets}
