% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{test.pairwise}
\alias{test.pairwise}
\title{Ouput of groupComparisonTMTPTM for full pairwise test}
\format{
A list of three dataframes
}
\usage{
test.pairwise
}
\description{
Returns the a list with three dataframes for three
statistical models. One for each Protein, PTM,
and PTM adjusted for protein level.
}
\details{
\itemize{
\item List objects: PTM.Model, Protein.Model, Adjusted.Model (all \code{dataframe}). Columns as follows:
\item Protein : Protein ID
\item Label: Label of the pairwise comparision or contrast
\item log2FC: Log2 fold change
\item SE: Standard error of the comparsion of contrast results
\item DF: Degree of freedom
\item pvalue: Value of p statistic of the test
\item adj.pvalue: adjusted p value
\item issue: used for indicating the reason why a comparison is not testable. NA means the comparison is testable.
'oneConditionMissing' means the protein has no measurements in one conndition of the comparison.
Furtherone, when 'issue = oneConditionMissing', 'log2FC = Inf' means the negative condition
(with coefficient -1 in the Label column)  is missing and 'log2FC = -Inf' means
the positive condition (with coefficient 1 in the Label column)  is missing.
completeMissing' means the protein has no measurements in all the connditions of the comparison.
unfittableModel' means there is no enough measurements to fit the linear model.
In other words, each condition has only one measurement.
}
}
\examples{
names(test.pairwise)
head(test.pairwise[[1]])

}
\keyword{datasets}
