% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMTPTM.R
\name{groupComparisonTMTPTM}
\alias{groupComparisonTMTPTM}
\title{Model PTM and/or protein data and make adjustments if needed}
\usage{
groupComparisonTMTPTM(
  data.ptm,
  data.protein = NULL,
  contrast.matrix = "pairwise",
  moderated = FALSE,
  adj.method = "BH"
)
}
\arguments{
\item{data.ptm}{Name of the output of proteinSummarization function with PTM data. It should have columns named
\code{Protein}, \code{TechRepMixture},  \code{Mixture}, \code{Run}, \code{Channel}, \code{Condition}, \code{BioReplicate}, \code{Abundance}.}

\item{data.protein}{Protein dataset returned by the proteinSummarization function}

\item{contrast.matrix}{Comparison between conditions of interests.
1) default is 'pairwise', which compare all possible pairs between two conditions.
2) Otherwise, users can specify the comparisons of interest. Based on the levels of conditions,
specify 1 or -1 to the conditions of interests and 0 otherwise.
The levels of conditions are sorted alphabetically.}

\item{moderated}{TRUE will moderate t statistic; FALSE (default) uses ordinary t statistic.}

\item{adj.method}{Adjusted method for multiple comparison. "BH" is default.}
}
\value{
A list \code{models} of all modeled and adjusted datasets
}
\description{
Takes summarized PTM data from proteinSummarization and models with
groupComparisonTMT. Can also take protein level data in the same format
and model with groupComparisonTMT. Including protein data allows
for adjusting PTM Fold Change by the change in protein abundance
without modification.
}
\examples{
# Load summarized datasets from MSstatsTMT proteinSummarization function
data(quant.msstats.ptm)
data(quant.msstats.protein)

# Load specific contrast matrix
data(example.comparisons)

test.pairwise <- groupComparisonTMTPTM(data.ptm=quant.msstats.ptm,
                                      data.protein=quant.msstats.protein,
                                      contrast.matrix = example.comparisons)

}
